requirements.txt
langchain
langchain_openai
openai
streamlit
python-dotenv
taskipy
